requirements.txt
torch
diffusers
transformers
accelerate
safetensors
gradio
python-dotenv